Bug #585: Edit AL - internal server error
git-svn-id: https://build.archivesportaleurope.net/svn/apex/branches/backend-project-1.3@1208 9fd842fd-667d-44da-918d-ff91420b3ae1

+++ Dashboard/src/main/java/eu/apenet/dashboard/manual/eag/WebFormEAG2012Action.java
@@ -804,10 +804,12 @@ public class WebFormEAG2012Action extends AbstractInstitutionAction {
 
 					// Move temp file to final file.
 					File eagFinalFile = new File((APEnetUtilities.getConfig().getRepoDirPath() + path));
-					try {
-						FileUtils.forceDelete(eagFinalFile);
-					} catch (IOException e) {
-						log.error(e.getMessage(),e);
+					if (eagFinalFile.exists()) {
+						try {
+							FileUtils.forceDelete(eagFinalFile);
+						} catch (IOException e) {
+							log.error(e.getMessage(),e);
+						}
 					}
 					FileUtils.moveFile(eagTempFile, eagFinalFile);
